Change over to more QtCore usage
Also working on gen 3/4 generator and searcher code.

Just wanted to get this committed before I accidentally deleted something.

+++ Objects/Frame.cpp
@@ -45,13 +45,13 @@ QString Frame::getGenderString()
 // Returns string equivalent of nature
 QString Frame::getNatureString()
 {
-    return natures[nature];
+    return natures[static_cast<int>(nature)];
 }
 
 // Returns string equivalent of hidden power
 QString Frame::getPowerString()
 {
-    return powers[hidden];
+    return powers[static_cast<int>(hidden)];
 }
 
 // Returns !!! if shiny or blank if not
